Moved funding tracker stub code to new location
